tests for migration
