+++ chains/orchestrator-relays/runtime/dancelight/src/tests/migrations_test.rs
@@ -15,13 +15,20 @@
 // along with Tanssi.  If not, see <http://www.gnu.org/licenses/>
 
 use crate::tests::common::ExtBuilder;
-use crate::{BeefyMmrLeaf, ExternalValidators, PalletInfo, Runtime, Session};
+use crate::{
+    BeefyMmrLeaf, CollatorConfiguration, ExternalValidators, PalletInfo, Runtime, Session,
+};
 use beefy_primitives::mmr::BeefyAuthoritySet;
 use frame_support::migration::clear_storage_prefix;
 use frame_support::storage::unhashed;
 use frame_support::traits::PalletInfo as _;
 use pallet_migrations::Migration;
-use tanssi_runtime_common::migrations::{ExternalValidatorsInitialMigration, MigrateMMRLeafPallet};
+use parity_scale_codec::Encode;
+use sp_arithmetic::Perbill;
+use tanssi_runtime_common::migrations::{
+    BondedErasTimestampMigration, ExternalValidatorsInitialMigration, HostConfigurationV3,
+    MigrateConfigurationAddFullRotationMode, MigrateMMRLeafPallet,
+};
 use xcm::v3::Weight;
 
 #[test]
@@ -113,5 +120,145 @@ fn test_migration_external_validators_pallet() {
         // ExternalValidators should be equal to validators from Session::queued_keys
         let expected_validators: Vec<_> = Session::queued_keys().into_iter().map(|x| x.0).collect();
         assert_eq!(migrated_validators, expected_validators);
+    });
+}
+
+#[test]
+fn test_migration_config_add_full_rotation_mode() {
+    ExtBuilder::default().build().execute_with(|| {
+        const CONFIGURATION_ACTIVE_CONFIG_KEY: &[u8] =
+            &hex_literal::hex!("86e86c1d728ee2b18f76dd0e04d96cdbb4b49d95320d9021994c850f25b8e385");
+        const CONFIGURATION_PENDING_CONFIGS_KEY: &[u8] =
+            &hex_literal::hex!("86e86c1d728ee2b18f76dd0e04d96cdb53b4123b2e186e07fb7bad5dda5f55c0");
+
+        // Modify active config
+        frame_support::storage::unhashed::put_raw(
+            CONFIGURATION_ACTIVE_CONFIG_KEY,
+            &HostConfigurationV3 {
+                max_collators: 5,
+                min_orchestrator_collators: 2,
+                max_orchestrator_collators: 1,
+                collators_per_container: 3,
+                full_rotation_period: 4,
+                collators_per_parathread: 2,
+                parathreads_per_collator: 1,
+                target_container_chain_fullness: Perbill::from_percent(45),
+                max_parachain_cores_percentage: Some(Perbill::from_percent(75)),
+            }
+            .encode(),
+        );
+        // Modify pending configs
+        frame_support::storage::unhashed::put_raw(
+            CONFIGURATION_PENDING_CONFIGS_KEY,
+            &vec![
+                (
+                    1234u32,
+                    HostConfigurationV3 {
+                        max_collators: 1,
+                        min_orchestrator_collators: 4,
+                        max_orchestrator_collators: 45,
+                        collators_per_container: 5,
+                        full_rotation_period: 1,
+                        collators_per_parathread: 1,
+                        parathreads_per_collator: 1,
+                        target_container_chain_fullness: Perbill::from_percent(65),
+                        max_parachain_cores_percentage: Some(Perbill::from_percent(75)),
+                    },
+                ),
+                (
+                    5678u32,
+                    HostConfigurationV3 {
+                        max_collators: 1,
+                        min_orchestrator_collators: 4,
+                        max_orchestrator_collators: 45,
+                        collators_per_container: 5,
+                        full_rotation_period: 1,
+                        collators_per_parathread: 1,
+                        parathreads_per_collator: 1,
+                        target_container_chain_fullness: Perbill::from_percent(65),
+                        max_parachain_cores_percentage: Some(Perbill::from_percent(75)),
+                    },
+                ),
+            ]
+            .encode(),
+        );
+
+        let migration = MigrateConfigurationAddFullRotationMode::<Runtime>(Default::default());
+        migration.migrate(Default::default());
+
+        let expected_active = pallet_configuration::HostConfiguration {
+            max_collators: 5,
+            min_orchestrator_collators: 2,
+            max_orchestrator_collators: 1,
+            collators_per_container: 3,
+            full_rotation_period: 4,
+            collators_per_parathread: 2,
+            parathreads_per_collator: 1,
+            target_container_chain_fullness: Perbill::from_percent(45),
+            max_parachain_cores_percentage: Some(Perbill::from_percent(75)),
+            ..Default::default()
+        };
+        assert_eq!(CollatorConfiguration::config(), expected_active);
+
+        let expected_pending = vec![
+            (
+                1234u32,
+                pallet_configuration::HostConfiguration {
+                    max_collators: 1,
+                    min_orchestrator_collators: 4,
+                    max_orchestrator_collators: 45,
+                    collators_per_container: 5,
+                    full_rotation_period: 1,
+                    collators_per_parathread: 1,
+                    parathreads_per_collator: 1,
+                    target_container_chain_fullness: Perbill::from_percent(65),
+                    max_parachain_cores_percentage: Some(Perbill::from_percent(75)),
+                    ..Default::default()
+                },
+            ),
+            (
+                5678u32,
+                pallet_configuration::HostConfiguration {
+                    max_collators: 1,
+                    min_orchestrator_collators: 4,
+                    max_orchestrator_collators: 45,
+                    collators_per_container: 5,
+                    full_rotation_period: 1,
+                    collators_per_parathread: 1,
+                    parathreads_per_collator: 1,
+                    target_container_chain_fullness: Perbill::from_percent(65),
+                    max_parachain_cores_percentage: Some(Perbill::from_percent(75)),
+                    ..Default::default()
+                },
+            ),
+        ];
+        assert_eq!(CollatorConfiguration::pending_configs(), expected_pending);
+    });
+}
+
+#[test]
+fn test_add_timestamp_to_bonded_eras_migration() {
+    ExtBuilder::default().build().execute_with(|| {
+        let bonded_eras_key =
+            pallet_external_validator_slashes::BondedEras::<Runtime>::hashed_key();
+
+        let previous_value: Vec<(sp_staking::EraIndex, sp_staking::SessionIndex)> =
+            vec![(1, 1), (2, 2), (3, 3)];
+
+        // Modify storage to pevious value
+        frame_support::storage::unhashed::put_raw(&bonded_eras_key, &previous_value.encode());
+
+        let migration = BondedErasTimestampMigration::<Runtime>(Default::default());
+        migration.migrate(Default::default());
+
+        let expected_bonded_eras_after: Vec<(sp_staking::EraIndex, sp_staking::SessionIndex, u64)> =
+            previous_value
+                .iter()
+                .map(|(era, session)| (*era, *session, 0u64))
+                .collect();
+        assert_eq!(
+            pallet_external_validator_slashes::BondedEras::<Runtime>::get(),
+            expected_bonded_eras_after
+        );
     });
 }
